fix: do no skip empty cell
