+++ src/train/src/prepare/mod.rs
@@ -3,6 +3,7 @@ mod parse;
 
 use std::collections::HashMap;
 
+use analyze::Analysis;
 use float::Float;
 use polars::prelude::*;
 
@@ -19,9 +20,23 @@ pub type Dataset = Vec<Features>;
 pub type Features = Vec<Float>;
 
 pub fn prepare(args: &Args, df: DataFrame) -> GroupedDatasets {
-	let mut grouped_datasets = parse::datasets(&df);
+	let analysis = features_analysis(&df);
 
-	// normalize::normalize(args, &df, &mut grouped_datasets);
+	let mut grouped_datasets = parse::datasets(&df, &analysis);
+
+	normalize::normalize(args, &mut grouped_datasets, &analysis);
 
 	grouped_datasets
+}
+
+fn features_analysis(df: &DataFrame) -> Vec<Analysis> {
+	let mut features_analysis = Vec::new();
+
+	for col in df.get_columns() {
+		if col.dtype().is_float() {
+			features_analysis.push(Analysis::from(col));
+		}
+	}
+
+	features_analysis
 }
